dev-requirements.txt
#
# This file is autogenerated by pip-compile with Python 3.13
# by the following command:
#
#    pip-compile --output-file=dev-requirements.txt dev-requirements.in
#
amqp==5.3.1
    # via kombu
asgiref==3.11.0
    # via
    #   -r requirements.in
    #   django
    #   django-cors-headers
billiard==4.2.4
    # via celery
black==25.11.0
    # via -r dev-requirements.in
celery==5.6.0
    # via
    #   -r requirements.in
    #   django-celery-beat
certifi==2025.11.12
    # via requests
charset-normalizer==3.4.4
    # via requests
click==8.3.1
    # via
    #   black
    #   celery
    #   click-didyoumean
    #   click-plugins
    #   click-repl
click-didyoumean==0.3.1
    # via celery
click-plugins==1.1.1.2
    # via celery
click-repl==0.3.0
    # via celery
colorama==0.4.6
    # via click
cron-descriptor==2.0.6
    # via django-celery-beat
django==5.2.9
    # via
    #   -r requirements.in
    #   django-celery-beat
    #   django-cors-headers
    #   django-filter
    #   django-timezone-field
    #   djangorestframework
    #   djangorestframework-simplejwt
    #   drf-yasg
django-celery-beat==2.8.1
    # via -r requirements.in
django-cors-headers==4.9.0
    # via -r requirements.in
django-filter==25.2
    # via -r requirements.in
django-timezone-field==7.2.1
    # via django-celery-beat
djangorestframework==3.16.1
    # via
    #   -r requirements.in
    #   djangorestframework-simplejwt
    #   drf-yasg
djangorestframework-simplejwt==5.5.1
    # via -r requirements.in
dnspython==2.8.0
    # via eventlet
drf-yasg==1.21.11
    # via -r requirements.in
eventlet==0.40.4
    # via -r requirements.in
exceptiongroup==1.3.1
    # via celery
greenlet==3.3.0
    # via eventlet
idna==3.11
    # via requests
inflection==0.5.1
    # via drf-yasg
kombu==5.6.1
    # via celery
mypy-extensions==1.1.0
    # via black
packaging==25.0
    # via
    #   black
    #   drf-yasg
    #   kombu
pathspec==0.12.1
    # via black
pillow==12.0.0
    # via -r requirements.in
platformdirs==4.5.1
    # via black
prompt-toolkit==3.0.52
    # via click-repl
psycopg2-binary==2.9.11
    # via -r requirements.in
pyjwt==2.10.1
    # via djangorestframework-simplejwt
python-crontab==3.3.0
    # via django-celery-beat
python-dateutil==2.9.0.post0
    # via celery
python-dotenv==1.2.1
    # via -r requirements.in
pytokens==0.3.0
    # via black
pytz==2025.2
    # via drf-yasg
pyyaml==6.0.3
    # via drf-yasg
redis==7.1.0
    # via -r requirements.in
requests==2.32.5
    # via
    #   -r requirements.in
    #   stripe
six==1.17.0
    # via python-dateutil
sqlparse==0.5.4
    # via
    #   -r dev-requirements.in
    #   django
stripe==14.0.1
    # via -r requirements.in
typing-extensions==4.15.0
    # via
    #   cron-descriptor
    #   stripe
tzdata==2025.2
    # via
    #   -r requirements.in
    #   django
    #   django-celery-beat
    #   kombu
    #   tzlocal
tzlocal==5.3.1
    # via celery
uritemplate==4.2.0
    # via drf-yasg
urllib3==2.6.0
    # via requests
vine==5.1.0
    # via
    #   amqp
    #   celery
    #   kombu
wcwidth==0.2.14
    # via prompt-toolkit
